Move interactive tools under /experimental routes
