+++ src/app/experimental/learning-dynamics/_components/MetricsPanel.tsx
