+++ src/app/experimental/mandelbrot/_components/MandelbrotCanvas.tsx
