+++ src/app/experimental/mandelbrot/_components/MandelbrotExplorer.tsx
